chore: refactor login page, add login form
